Add a FlagSpec registry for node and edge flags
Node and edge flags were a fixed catalog of compile-time bits with no
name table, so a plugin could only emit the built-ins and an external
reader of the graph file couldn't map a bit back to a name. Introduce a
single FlagSpec mechanism with two registration paths: the engine
registers its engine/* flags with explicit masks (hot-path reads still
constant-fold), and plugins declare flags via the new
declare_node_flags()/declare_edge_flags() hooks, with the host allocating
each a bit above the engine masks in registration order (deterministic).
The engine/ namespace is reserved, identical specs registered by two
plugins share one bit (pytest and unittest now share test/testcase), and
a conflicting re-declaration or exhausting the bit width fails loudly.

Both registries are serialized into the .dcg file (FORMAT_VERSION 1->2,
old files rejected) so an external explorer can decode any bit by name.
Node flags stay u32; edge flags narrow to u8.

Roster surgery: drop the dead SHADOWED and EXPORTED flags; move OVERLOAD
off the flag space into an internal build-time is_overload field on the
node payload (not serialized, not Python-visible); move TESTCASE to the
shared test/testcase plugin flag; add an engine dead_branch node flag.
The hand-maintained KEEPALIVE_DEFAULT constant is replaced by the
registry-derived default_seed_mask() (OR of seed && default_on bits), so
test/testcase is kept alive by default iff a test plugin is registered.

PLUGIN_API_EPOCH bumped 1->2 (the curated plugin_api surface changed).

+++ runtime/src/file_ref_edges.rs
@@ -91,7 +91,7 @@ use crate::ingest::{
 /// GIL-free so workers run inside `py.allow_threads` cleanly.
 #[derive(Debug, Eq, PartialEq, salsa::Update, get_size2::GetSize)]
 pub(crate) struct FileRefEdges<'db> {
-    pub(crate) edges: FxHashSet<(NodeRef<'db>, NodeRef<'db>, u32)>,
+    pub(crate) edges: FxHashSet<(NodeRef<'db>, NodeRef<'db>, u8)>,
     pub(crate) warnings: Vec<String>,
 }
 
@@ -101,7 +101,7 @@ pub(crate) struct FileRefEdges<'db> {
 #[salsa::tracked(returns(ref), heap_size = ruff_memory_usage::heap_size)]
 pub(crate) fn file_to_ref_edges<'db>(db: &'db dyn ProjectDb, file: File) -> FileRefEdges<'db> {
     let self_nodes = file_to_nodes(db, file);
-    let mut edges: FxHashSet<(NodeRef<'db>, NodeRef<'db>, u32)> = FxHashSet::default();
+    let mut edges: FxHashSet<(NodeRef<'db>, NodeRef<'db>, u8)> = FxHashSet::default();
     let mut warnings: Vec<String> = Vec::new();
 
     let parsed = parsed_module(db, file).load(db);
@@ -204,7 +204,7 @@ struct RefWalker<'a, 'db> {
     /// `TYPE_CHECKING` as `True`); `find_local_bindings` recovers it
     /// from the scope-wide reachable bindings.
     tc_ranges: &'a [TextRange],
-    edges: &'a mut FxHashSet<(NodeRef<'db>, NodeRef<'db>, u32)>,
+    edges: &'a mut FxHashSet<(NodeRef<'db>, NodeRef<'db>, u8)>,
     /// Per-file warnings buffer. Workers push pure-rust strings; the
     /// driver flushes them to Python logging from the main thread.
     warnings: &'a mut Vec<String>,
@@ -225,7 +225,7 @@ struct RefWalker<'a, 'db> {
     /// (set by `emit_name_use` / nested-import handlers from
     /// `flags_for_range` on the reference's source position;
     /// reset to 0 after the reference completes).
-    current_flags: u32,
+    current_flags: u8,
 }
 
 impl<'a, 'db> RefWalker<'a, 'db> {
@@ -237,7 +237,7 @@ impl<'a, 'db> RefWalker<'a, 'db> {
 
     /// Returns `EDGE_FLAG_DEAD_BRANCH` if `range` is contained in any
     /// statically-dead region recorded for this file, else `0`.
-    fn flags_for_range(&self, range: TextRange) -> u32 {
+    fn flags_for_range(&self, range: TextRange) -> u8 {
         if self.dead_ranges.iter().any(|r| r.contains_range(range)) {
             EDGE_FLAG_DEAD_BRANCH
         } else {
